HealthDoc-AI/requirements.txt
streamlit==1.31.1
langchain==0.1.9
openai==1.12.0
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.6.0
faiss-cpu==1.7.4
langchain-openai==0.0.8
langchain-community==0.0.24
sentence-transformers==2.5.1
python-dotenv==1.0.1